finish task 3.1

+++ src/lcmtypes/occupancy_grid_t.hpp
@@ -75,13 +75,13 @@ class occupancy_grid_t
         inline int _encodeNoHash(void *buf, int offset, int maxlen) const;
         inline int _getEncodedSizeNoHash() const;
         inline int _decodeNoHash(const void *buf, int offset, int maxlen);
-        inline static int64_t _computeHash(const __lcm_hash_ptr *p);
+        inline static uint64_t _computeHash(const __lcm_hash_ptr *p);
 };
 
 int occupancy_grid_t::encode(void *buf, int offset, int maxlen) const
 {
     int pos = 0, tlen;
-    int64_t hash = getHash();
+    int64_t hash = (int64_t)getHash();
 
     tlen = __int64_t_encode_array(buf, offset + pos, maxlen - pos, &hash, 1);
     if(tlen < 0) return tlen; else pos += tlen;
@@ -204,9 +204,9 @@ int occupancy_grid_t::_getEncodedSizeNoHash() const
     return enc_size;
 }
 
-int64_t occupancy_grid_t::_computeHash(const __lcm_hash_ptr *)
+uint64_t occupancy_grid_t::_computeHash(const __lcm_hash_ptr *)
 {
-    int64_t hash = 0x6aa23f24a5336649LL;
+    uint64_t hash = 0x6aa23f24a5336649LL;
     return (hash<<1) + ((hash>>63)&1);
 }
