+++ src/lcmtypes/particle_t.hpp
@@ -71,13 +71,13 @@ class particle_t
         inline int _encodeNoHash(void *buf, int offset, int maxlen) const;
         inline int _getEncodedSizeNoHash() const;
         inline int _decodeNoHash(const void *buf, int offset, int maxlen);
-        inline static int64_t _computeHash(const __lcm_hash_ptr *p);
+        inline static uint64_t _computeHash(const __lcm_hash_ptr *p);
 };
 
 int particle_t::encode(void *buf, int offset, int maxlen) const
 {
     int pos = 0, tlen;
-    int64_t hash = getHash();
+    int64_t hash = (int64_t)getHash();
 
     tlen = __int64_t_encode_array(buf, offset + pos, maxlen - pos, &hash, 1);
     if(tlen < 0) return tlen; else pos += tlen;
@@ -160,7 +160,7 @@ int particle_t::_getEncodedSizeNoHash() const
     return enc_size;
 }
 
-int64_t particle_t::_computeHash(const __lcm_hash_ptr *p)
+uint64_t particle_t::_computeHash(const __lcm_hash_ptr *p)
 {
     const __lcm_hash_ptr *fp;
     for(fp = p; fp != NULL; fp = fp->parent)
@@ -168,7 +168,7 @@ int64_t particle_t::_computeHash(const __lcm_hash_ptr *p)
             return 0;
     const __lcm_hash_ptr cp = { p, (void*)particle_t::getHash };
 
-    int64_t hash = 0x18016676f01f14e8LL +
+    uint64_t hash = 0x18016676f01f14e8LL +
          pose_xyt_t::_computeHash(&cp) +
          pose_xyt_t::_computeHash(&cp);
