+++ src/lcmtypes/pose_xyt_t.hpp
@@ -66,13 +66,13 @@ class pose_xyt_t
         inline int _encodeNoHash(void *buf, int offset, int maxlen) const;
         inline int _getEncodedSizeNoHash() const;
         inline int _decodeNoHash(const void *buf, int offset, int maxlen);
-        inline static int64_t _computeHash(const __lcm_hash_ptr *p);
+        inline static uint64_t _computeHash(const __lcm_hash_ptr *p);
 };
 
 int pose_xyt_t::encode(void *buf, int offset, int maxlen) const
 {
     int pos = 0, tlen;
-    int64_t hash = getHash();
+    int64_t hash = (int64_t)getHash();
 
     tlen = __int64_t_encode_array(buf, offset + pos, maxlen - pos, &hash, 1);
     if(tlen < 0) return tlen; else pos += tlen;
@@ -162,9 +162,9 @@ int pose_xyt_t::_getEncodedSizeNoHash() const
     return enc_size;
 }
 
-int64_t pose_xyt_t::_computeHash(const __lcm_hash_ptr *)
+uint64_t pose_xyt_t::_computeHash(const __lcm_hash_ptr *)
 {
-    int64_t hash = 0x0f98bd7892313b56LL;
+    uint64_t hash = 0x0f98bd7892313b56LL;
     return (hash<<1) + ((hash>>63)&1);
 }
